fix bugs when testing on windows with python3.7

+++ openbox/surrogate/base/build_gp.py
@@ -15,8 +15,8 @@ def create_gp_model(model_type, config_space, types, bounds, rng):
         prior=LognormalPrior(mean=0.0, sigma=1.0, rng=rng),
     )
 
-    cont_dims = np.nonzero(types == 0)[0]
-    cat_dims = np.nonzero(types != 0)[0]
+    cont_dims = np.nonzero(types == 0)[0].astype(np.int)
+    cat_dims = np.nonzero(types != 0)[0].astype(np.int)
 
     if len(cont_dims) > 0:
         exp_kernel = Matern(
